# openapi.yaml
openapi: 3.0.0
info:
  title: saborlocal API
  description: API REST completa con autenticación JWT y CRUD de entidades
  version: 1.0.0
servers:
  - url: http://localhost:3008/api
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EntityBase:
      type: object
      properties:
        _id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        imagen:
          type: string
        imagenThumbnail:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
paths:
...
  /productor:
    get:
      tags:
        - Productor
      summary: Listar todos los Productors
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Productors
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Productor
      summary: Crear Productor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Productor creado
  /productor/{id}:
    get:
      tags:
        - Productor
      summary: Obtener Productor por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Productor encontrado
        '404':
          description: Productor no encontrado
    patch:
      tags:
        - Productor
      summary: Actualizar Productor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Productor actualizado
    delete:
      tags:
        - Productor
      summary: Eliminar Productor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Productor eliminado
  /productor/{id}/upload-image:
    post:
      tags:
        - Productor
      summary: Subir imagen a Productor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
  /producto:
    get:
      tags:
        - Producto
      summary: Listar todos los Productos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Productos
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Producto
      summary: Crear Producto
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Producto creado
  /producto/{id}:
    get:
      tags:
        - Producto
      summary: Obtener Producto por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Producto encontrado
        '404':
          description: Producto no encontrado
    patch:
      tags:
        - Producto
      summary: Actualizar Producto
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Producto actualizado
    delete:
      tags:
        - Producto
      summary: Eliminar Producto
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Producto eliminado
  /producto/{id}/upload-image:
    post:
      tags:
        - Producto
      summary: Subir imagen a Producto
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
...